+++ TUM Campus App/CalendarManager.swift
@@ -9,11 +9,12 @@
 import Foundation
 import Sweeft
 
-final class CalendarManager: CachedManager, CardManager, SingleItemCachedManager {
+final class CalendarManager: MemoryCachedManager, CardManager, SingleItemCachedManager {
     
     typealias DataType = CalendarRow
     
     var config: Config
+    var cache: Cache<[CalendarRow]>?
     
     var defaultMaxCache: CacheTime {
         return .time(.aboutOneWeek)
@@ -32,13 +33,13 @@ final class CalendarManager: CachedManager, CardManager, SingleItemCachedManager
     }
     
     func toSingle(from items: [CalendarRow]) -> DataElement? {
-        return items.first { $0.start > .now }
+        return items.first { $0.irrelevantAfter > .now && $0.status == "FT" }
     }
     
-    func fetch(maxCache: CacheTime) -> Response<[CalendarRow]> {
+    func performRequest(maxCache: CacheTime) -> Response<[CalendarRow]> {
         return config.tumOnline.doXMLObjectsRequest(to: .calendar,
                                                     at: "events", "event",
-                                                    maxCacheTime: maxCache).map { $0.filter { $0.status == "FT" } }
+                                                    maxCacheTime: maxCache)
     }
     
 }