+++ TUM Campus App/CalendarRow.swift
@@ -64,7 +64,7 @@ final class CalendarRow: DataElement {
 
 extension CalendarRow: XMLDeserializable {
     
-    convenience init?(from xml: XMLIndexer) {
+    convenience init?(from xml: XMLIndexer, api: TUMOnlineAPI, maxCache: CacheTime) {
         
         guard let title = xml["title"].element?.text,
             let start = xml["dtstart"].element?.text.date(using: "yyyy-MM-dd HH:mm:ss"),
@@ -80,6 +80,14 @@ extension CalendarRow: XMLDeserializable {
                   status: status,
                   location: xml["location"].element?.text,
                   url: xml["url"].element?.text.url)
+    }
+    
+}
+
+extension CalendarRow {
+    
+    var irrelevantAfter: Date {
+        return min(start.addingTimeInterval(30 * .oneMinute), end.addingTimeInterval(-15 * .oneMinute))
     }
     
 }