+++ TUM Campus App/GradeTableViewCell.swift
@@ -14,13 +14,12 @@ class GradeTableViewCell: CardTableViewCell {
         didSet {
             if let grade = grade {
                 titleLabel.text = grade.name
-                resultLabel.text  = "Restult: " + grade.result
+                resultLabel.text  = "Result: " + grade.result
                 let date = "Date: \(grade.date.string(using: "dd.MM.yyyy"))"
                 let semester = "Semester: \(grade.semester)"
-                let ects = "ECTS: \(grade.ects)"
                 let examiner = "Examiner: \(grade.examiner)"
                 let mode = "Mode: \(grade.mode)"
-                detailsLabel.text =  date + ", " + semester +  ", " + ects
+                detailsLabel.text =  date + ", " + semester
                 secondDetailsLabel.text =  examiner + ", " + mode
             }
         }