+++ TUM Campus App/NewsManager.swift
@@ -9,11 +9,12 @@
 import Foundation
 import Sweeft
 
-final class NewsManager: CachedManager, SingleItemCachedManager, CardManager {
+final class NewsManager: MemoryCachedManager, SingleItemCachedManager, CardManager {
     
     typealias DataType = News
     
     var config: Config
+    var cache: Cache<[News]>?
     
     var requiresLogin: Bool {
         return false
@@ -35,9 +36,22 @@ final class NewsManager: CachedManager, SingleItemCachedManager, CardManager {
         return items.filter({ $0.date > .now }).last ?? items.first
     }
     
-    func fetch(maxCache: CacheTime) -> Response<[News]> {
+    private func fetch(from news: String) -> Response<[News]> {
         return config.tumCabe.doObjectsRequest(to: .news,
-                                               maxCacheTime: maxCache).map { $0.sorted(descending: \.date) }
+                                               arguments: ["news" : news],
+                                               maxCacheTime: .forever).flatMap { (results: [News]) in
+            
+            guard let lastId = results.max({ $0.id }) else {
+                self.config.tumCabe.removeCache(for: .news, arguments: ["news" : news])
+                return .successful(with: results)
+            }
+            return self.fetch(from: lastId).map { results + $0 }
+                                           .mapError(to: results)
+        }
+    }
+    
+    func performRequest(maxCache: CacheTime) -> Promise<[News], APIError> {
+        return fetch(from: "").map { $0.sorted(descending: \.date) }
     }
     
 }