+++ TUM Campus App/PersonSearchManager.swift
@@ -30,12 +30,15 @@ final class PersonSearchManager: SearchManager {
         self.config = config
     }
     
+    func search(query: String, maxCache: CacheTime) -> Response<[UserData]> {
+        return config.tumOnline.doXMLObjectsRequest(to: .personSearch,
+                                                    queries: ["pSuche" : query],
+                                                    at: "rowset", "row",
+                                                    maxCacheTime: maxCache)
+    }
+    
     func search(query: String) -> Promise<[UserData], APIError> {
-        return config.tumOnline.doRepresentedRequest(to: .personSearch,
-                                                     queries: ["pSuche" : query]).map { (xml: XMLIndexer) in
-            
-            return xml.get(at: ["rowset", "row"])?.all ==> { UserData(from: $0, api: self.config.tumOnline) }
-        }
+        return search(query: query, maxCache: .no)
     }
     
 }