+++ TUM Campus App/TUMSexyAPI.swift
@@ -24,4 +24,11 @@ struct TUMSexyAPI: API {
     
     let baseURL: String
     
+    
+    var baseHeaders: [String : String] {
+        return [
+            "User-Agent": Bundle.main.userAgent,
+        ]
+    }
+    
 }