+++ CHANGELOG.md
@@ -1,4 +1,24 @@
-## [1.3.0] - Unreleased
+## [1.3.1] - Unreleased
+### Added
+- Added User Profile Picture to "More" Button.
+
+### Changed
+- Current User Information and Profile Image is now being cached as well.
+- Pinning Campus App API Root CA Certificate for improved security.
+- TUM Online Token Name now includes device nickname for identification convenience.
+- Updated User-Agent Header now displays the App Version and Build Number
+- Only displaying Tuition Status in Card View when not payed.
+- Changed up the App Icon with a bigger TUM Logo. True excellence has been achieved ;)
+
+### Fixed
+- Fixed displaying no user information for the current user in case the name is shared by multiple people.
+- Fixed Typo "Restults" from the grades view.
+
+### Removed
+- Removed faulty ECTS label from grades.
+
+
+## [1.3.0] - 2018-01-24
 ### Added
 - Added 💸 Ca$hing 💸 for better performance
 - Added Headers to Search to mark the type of results
@@ -9,12 +29,14 @@
 - Links now open within the App to avoid annoying app switches
 - If the app doesn't have permissions to access your location it will now assume you're at the campus.
 - Pulling to refresh on the Card View will now skip all caches.
+- Calendar now displays canceled events grayed out.
 
 ### Fixed
 - Fixed memory leaks caused by memory cycles.
 - Fixed Cafeteria Card. It only shows a Cafeteria that has something on the menu that day
 - Fixed Calendar View not changing the day after a few swipes.
 - Fixed More View displaying the wrong name.
+- Fixed TU Film Links
 
 ### Removed
 - Removed Tab Bar in favor of a Bar Button at the Top